requirements.txt
absl-py
attrs
dataframe_image
einops
idna
imageio
matplotlib
mpyq
numpy
pbr
Pillow
pluggy
portpicker
probscale
protobuf
py
pygame
pyparsing
pytablewriter
pytest
python-dateutil
PyYAML==5.3.1
requests
sacred
scipy
seaborn
six
sk-video
snakeviz
# torch-scatter # required, but will depend on your torch version 
tensorboard-logger
tornado
urllib3
websocket-client
whichcraft
wrapt